requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
coverage==5.5
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
idna==2.10
importlib-metadata==3.10.1
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
MarkupSafe==1.1.1
numpy==1.20.2
packaging==20.9
pandas==1.2.4
pluggy==0.13.1
pprintpp==0.4.0
psycopg2==2.8.6
py==1.10.0
pycountry==20.7.3
pycountry-convert==0.7.2
pyparsing==2.4.7
pytest==6.2.3
pytest-cov==2.11.1
pytest-mock==3.5.1
python-dateutil==2.8.1
pytz==2021.1
pyweatherbit==2.0.0
repoze.lru==0.7
requests==2.25.1
responses==0.13.2
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
sklearn==0.0
SQLAlchemy==1.4.9
threadpoolctl==2.1.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
zipp==3.4.1
